FAB-14491 Move ccmetadata out of platforms
The ccmetadata package provides assorted sanity-checking and parsing for
chaincode metadata info.  It's called directly by lscc, and by the new
lifecycle, as well as in the sanity checking for platform deployment
spec checks.  It really has nothing to do with a chaincode platform, so,
moving it into internal/pkg for lack of a better place.

Change-Id: I7a9be54eafd17ceed80463000ec6c889b80ae274
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>