FAB-14491 Add basic sanitizing untar utility
For the external builders to work, we need to be able to expand code
packages into a tmp directory to run 'detect' against.  This CR adds a
basic util function for safely expanding tar files, ensuring that only
normal files and directories are created and as children of the
temporary source directory.

Change-Id: I51201288d515d506d49863d2f5ebce6b862cb549
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>