FAB-11954 Pass peer addr and TLS together
Presently, the docker controller is configured with the peer address of
the chaincode server that chaincodes should connect back to.  This is a
bit awkward, because the other connection into, the TLS parameters, come
across at start time.  Passing these two together simplifies the concept
and prepares us to pass the two together to 'launch' in the external
builder.

Change-Id: Ia550cfe493d57e79ee61a5ce0230deede252c0f6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>