FAB-11594 Output instances from build
Today, the container interface is stateless -- it allows you to make
invocations against containers under the assumption that someone has
come before and built the instance to start or stop.  This CR leaves the
high level stateless interface, but changes the internals to associate
the output of Build with a chaincode id.

This is a prerequisite for the chaincode launcher work which cannot be
stateless.

Change-Id: I3576493f17368a1b44e782b02da76a2356901f60
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>