[FAB-15897] Improve CLI perf w/ multiple endorsers
If you specify multiple peers to "peer chaincode invoke", currently
these endorsements are run one at a time, in order specified on the CLI.

This is slow, which might not be seen as a problem for the CLI, however
in the current version of chaincode lifecycle this command is used to
trigger the building of chaincode containers. This may be causing CI failures
when chaincode needs to be built in a single transaction across multiple
peers.

This change runs the endorsements concurrently, which should trigger
the chaincode to be built across all peers simultaneously. Also, this is
how the various SDKs behave today.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: Ib9f69c7bbd7f4f1197f2959599e45d0a2545d366