FAB-14493 Push the awkward base64 encoding down
For historical reasons, (cert material used to be passed as a flag), the
cert information for chaincodes is currently encoded into base64 in the
cert files, for the client cert and key, but not for the root cert.
This is awkward, but is hidden by having the 'cert generator' provide
the certs opaquely as strings, which are then treated as bytes.

This CR simply pushes the encoding out to the periphery, closest to
where the file is written, so as to avoid even more awkward things, like
the chaincode.json having a base64 encoded version of a base64 encoded
version of a PEm encoded string.

Change-Id: I14490c4fa7d14382ff665eafbbcaa31355a5996a
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>