[FAB-16165] Change pkcs11 test keystore directory
PKCS11 tests currently use the system temp directory
as the location for the keystore. The PKCS11 package looks for
the keystore by looping over files in the directory specified,
reading them, and then attempting to convert them to PEM files.
If a file with no EOF marker, or an open pipe exists in the
directory, attempting to read it deadlocks the read operation
in the BCCSP library.

Instead we create a sub-directory in the temp directory
and create the keystore there

Signed-off-by: Brett Logan <Brett.T.Logan@ibm.com>
Change-Id: I007a7aad6efa921cf9dd6c44c9ae921c68a29db9
(cherry picked from commit e841ebe1b04063de26d3583de1a425cd77b7d844)