FAB-16166 Enforce new package file names
We are transitioning Code-Package-Metadata.json to be simply
metadata.json and the 'other file is always the code package', to be
'the code package must be named 'code.tar.gz'.

The latter change has already been merged into the SDKs and CLI.  For
now, we allow a fallback to the old metadata name, but this will be
removed once the SDKs transition.

Change-Id: I7690a77f63a11daaac4a25ffda52135e0958eb8b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>