[FAB-16184] fix make verify
run-unit-tests includes 'set -e -o pipefail' to make the script fail
when commands fail. When running with JOB_TYPE=VERIFY, the grep
invocation embedded in 'packages_diff' would fail when no go files were
changed and this caused the script to fail.

Once that was addressed, it exercised new paths related to package_spec
where the variable would contain a new line separated list of packages
instead of a single package spec. This was addressed by parsing the
result of package_diff into an array and treating package_spec strictly
as an array.

Change-Id: I66d07fd620deb0d08d2c22e50a1b36d5cdbf8f1f
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>