FAB-16124 Remove unnecessary dep on SysCCVersion
The validation code currently populates the 'SystemChaincodeVersion' of
the peer (ie, the peer's build level) as the default chaincode version,
and then overrides it if the chaincode is not a system chaincode.

However, the only place where the version of the chaincode is checked is
in the event that it is _not_ a system chaincode, so, populating this
value has no effect.

It's important to note that if this version string ever were used, it
would be hopelessly non-deterministic, so it is safest simply to remove
it.

Change-Id: Iacdd90b2c8a830463fde025e97524913d32cbcd8
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>