rollback/reset: ensure that stateDB is dropped
Currently, the peer node reset/rollback command itself
take care of dropping the statelevelDB. However, the
couchDB should be removed manually. As an admin could
do a mistake by not removing the couchDB, we need to
add necessary check in the code to ensure that the
stateDB is dropped after a reset/rollback.

FAB-16113 #done

Change-Id: I5f68371ec0bebb4ab35d65eba38033a4c97a27be
Signed-off-by: senthil <cendhu@gmail.com>