rollback/reset: rm sync up between pvtStr-blkStr
With the introduction of peer reset/rollback, we make the
gossip to fetch pvtData from the local store if available.
Further, we make the pvtdata reconciler to not fetch pvtData
when the pvtstore is ahead of the blockstore. As we allow
the pvtstore height to be greater than the blockstore height,
we can safely remove the synchronization between pvtstore
and blockstore.

FAB-16211 #done

Change-Id: I38a6350c09f7216ef06dbf07636e686bc00e8217
Signed-off-by: senthil <cendhu@gmail.com>