FAB-16213 remove shim/ext/entities
The ext/entities package has dependencies on bccsp that prevent its
extraction from fabric as a standalone library. As moving the shim out
of fabric is a priority, the package is being removed.

The example chaincode that demonstrated the use of this package was
removed from the fabric tree quite some time ago.

Change-Id: I0d6e602936ec3f33bd286abe196f7be8147af01b
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>