FAB-16158 Have definition include packageID
Today, the chaincode flow looks roughly like this:

1) endorser.go looks up chaincode definition and call core/chaincode
2) core/chaincode looks up the chaincode container info, calls container
3) core/chaincode looks up the chaincode definition, again

There are a lot of things wrong with this, but one obvious one is that
the container info has nothing to do with core/chaincode, and purely
from the chaincode definition it should be good to operate.

This CR adds the ChaincodeID, ie the unique handle by which the
core/chaincode package references running chaincodes, to the chaincode
definition as returned by the lifecycle implementation.  This paves the
way for only the container package to actually get the container info
and helps decouple the notion of package ID, and chaincode ID.

Change-Id: Idb2f2b32c8e416ab7c0774500368cedb311d1a92
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>