FAB-16158 Get chaincode package without CCCI dep
For legacy reasons which will be eliminated in a future CR, it's
necessary to retrieve the container info for chaincodes at every
invocation.  We then pass the container info down to the container
package, but this info is effectively read again when retreiving the
code package.  This is an intermediate step to eliminating the container
info retreival call.

Change-Id: Iad1a8507be6f87a352750348dfb6ea151180e63c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>