FAB-16353 Move proposal validation to endorser
This is the first CR in a series which simply cleans up the endorser.go
code.  The endorser package is the only consumer of the
core/common/validation.ValidateProposal code, and it's not at all clear
why this code was put into the command package instead put to live with
th endorser to begin with.

Note, this CR does delete quite a bit of (bad) tests.  This code as
copied also relies on evil things like the mspmgt singleton.  Because of
this, no tests have been added.  Instead, the code overall will be
refactored with good tests added towards the end.  Ideally this series
should merge all at once to avoid any transient gaps in code coverage.

Change-Id: Ibe31fd62b3c3a79b67e4639614fe8ffdf99c4c18
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>