FAB-16353 Stop pretending chaincode version is set
The client has no idea what version chaincode they are invoking, and
they never set it.  Yet all of our endorser tests seem to believe that
they do.  Further, we build all of our metrics based on the idea that
'chaincode-id' is a property of the invocation, when in fact, it is only
chaincode name that we have access to for many of the checks.

This results in metrics that look like "mycc:" being reported as the
"chaincode" label.

If we feel that version is important to be reported for the metric, or,
if we actually want to report the chaincode id, as used internally by
the chaincode subsystem (which in the legacy case is name:version or in
the new lifecycle case is label:hash), then we can.

Change-Id: Iadda35437ccb94a133692e0413ac4a87b936eaf8
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>