FAB-16353 Return an error rather than ProposalResp
Directly returning a proposal response that has an error message set and
a status of 500, embedding an error, over and over again through the
endorsement code is unweidly.  Better to work with raw errors and bubble
them up to the top.

Note, this CR causes the previously returned error (with no response) on
chaincode event marshaling errors to go back to the client.  There is no
obvious reason marshaling a chaincode event should ever fail, nor why we
would want to hide this error text from the client, so this seems like a
safe error.

Change-Id: I34c082e4aac406ba34749913e0c62349e27261ac
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>