FAB-16359 Remove configtxlator sanitycheck
The configtxlator sanitycheck function was a well intentioned bit of
function that never actually came to fruition.  It is not documented
anywhere, nor is it believed to be used by anyone.  As this creates a
sizable dependency on the fabric code for what is otherwise a
straightforward utility, and creates a maintenance burden, v2.0 seems
like a good time to remove it.

Change-Id: Ibddbd032c59c099c465dc540a7dc2726e39582ab
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>