Check HistoryDB format
This CR adds code to leveldbhelper to check whether the db is in the
specified format. If the db is not in the specified format, and it
contains some data, then an error is returned at the time of opening.
If the db is empty, a key is written to record the specified format
so that subsequent opening of the db succeeds.

This CR also uses the above feature for enabling
the format check for the historydb

FAB-16386 #done

Change-Id: If3e8ab2cd43ebae134618c78bcfbd65bd594d89a
Signed-off-by: manish <manish.sethi@gmail.com>