mv DB's loc. constructor code to a single place
Currently, we construct the path of various DBs and stores
in many places within the ledger component. This is risky
as we need to ensure that the same postfix is used during
a path construction at all places. Hence, in this CR, we
move this logic to a common place.

FAB-16126 #done

Change-Id: I04b24f85d64b5867200316add3cd6423cc3e8da3
Signed-off-by: senthil <cendhu@gmail.com>