FAB-7559 Don't round trip marshal blocks
While exploring how to nicely wire in the org specific endpoints into
the deliver service, I stumbled across this code.

In the standard case, when a block arrives from the Deliver service, we
marshal it to bytes, then pass those bytes to be verified, and then
immediately turn around and verify those bytes.  It's not at all clear
why an internal API was ever written to consume bytes instead of an
unmarshaled structure, but, this CR fixes it, and eliminates a
significant amount of wasted cycles and garbage.

Change-Id: Ibd5b06ab84c2db3f54e65f3f8ba9e79c3977ec69
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>