FAB-15732 Remove validation packaging dependency
For historical reasons, which are poor, the v12/v13/v14 validation
depends upon the platforms code for performing validity checks on
chaincode deployment specs.  This is a bad thing for a number of
reasons, notably, because it's not at all obvious that changing the
packaging checks should affect validation across all three plugins, but
further, because there should never be a code package on the ledger
anyway.

When the peer cli, or any of the SDKs performs an instantiate, the
'code_package' field of the chaincode deployment spec is nil-ed out.
This is because the code package has already been installed, and LSCC
ignores this code package if it is present, instead using the one on the
filesystem.  If the CLI and SDKs included the code package,
instantiation transactions would be huge and problematic.

Therefore, the code checks of ValidateDeploymentSpec are actually
irrelevant for the validation path.  The first check perfomed by
'ValidateDeploymentSpec' is to check if the platform exists, then, check
if the code package is nil.  If both are true, the validation returns
successfully.  Therefore, there is no need to check the whole packaging
path, instead, it's only needed to check that the chaincode type is
defined.

Change-Id: I8e8054d2341a66117fcaab14e0915ad0b5ffb6ae
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>