Index all txs without duplicate TxIDs check
In the current code, we index only one transaction (the first one) with
a txid. This is a rare scenario that we will encounter more than one
transaction with same txid. However, this requires all over the code
to handle this situation making the code complex.

This CR, encodes the txid such that the index key for a txid becomes
unique and can simply be added to the index without worrying about
actual duplicates. The encoding leverages the block number and
transaction number to achieve this.

During lookup, first such entry for a txid (in practice, there should be
only one such entry) is retuned which maintains the current semantics
(i.e., queries return only first transaction with a given txid).
However, the new index would be able to support another query that
returns all the transactions with a given query in future.

Because of above mentioned new encoding, this requires a rebuild of
block storage index on the peer

FAB-16189 #done

Change-Id: Iac3ea84fbb4c4228202def2c5275aa8c8f03b90a
Signed-off-by: manish <manish.sethi@gmail.com>