FAB-13528 Add basic build status registry
Because both the chaincode launch-on-invoke path and the
build-on-install path will need to be able to build a chaincode, some
form of synchronization is needed between them.  This CR creates a small
registry, following the 'context' model of a done channel and an error
set on the structure, to allow coordination between the two.  It does
not attempt to wire it yet.

Change-Id: If89135b1f2e40cf15c38ea481be198da992f61d0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>