[FAB-16560] []SignedData to []Identity converter
As part of the policy rework, we need a helper function to convert a slice of
signed data into a slice of identities, such that each identity in the slice
is valid according to the supplied identity deserialiser and it produced a
valid signature over the message. Furthermore, the converter deduplicates
identities.

Change-Id: I26f381024527b449b5d1567fe89e14a46fd61d48
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>