FAB-16605 Add log message for slow WAL
When calls to store data in the WAL take too long, bad things will
happen in the network.  Notably, if the leader takes close to, or more
than the election timeout value, then the network will elect a new
leader and end up discarding the transaction he's working on.

This CR adds a simple log warning statement, which is printed if the WAL
storage time exceeds half of the configured election timeout.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: Ide6fcbbbb83477f483bdebeec6f3953fe1f94085